chatbot/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
langchain==0.2.12
langchain-core==0.2.28
langchain-groq==0.1.6